modified update blog and delete blog

+++ src/app/modules/blog/blog.service.ts
@@ -1,14 +1,42 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
 import QueryBuilder from '../../builder/queryBuilder';
 import { blogSearchableFields } from './blog.constant';
 import { TBlog } from './blog.interface';
 import { Blog } from './blog.model';
+import { User } from '../user/user.model';
+import config from '../../config';
+import jwt, { JwtPayload } from 'jsonwebtoken';
+import AppError from '../../errors/AppError';
+import { StatusCodes } from 'http-status-codes';
 
 const createBlog = async (payload: TBlog) => {
-  const result = await Blog.create(payload);
+  const result = (await Blog.create(payload));
   return result;
 };
 
-const updateBlog = async (id: string, payload: TBlog) => {
+const updateBlog = async (id: string, payload: TBlog, token: string) => {
+  const data = await Blog.findById(id);
+
+  const authorId = data?.author;
+  const authorEmail = (await User.findById(authorId))?.email;
+  
+
+  if (!token) {
+    throw new AppError(StatusCodes.UNAUTHORIZED, 'You are not authorized');
+  }
+
+  const decoded = jwt.verify(
+    token,
+    config.jwt_access_secret as string,
+  ) as JwtPayload;
+
+  const { email } = decoded;
+
+  if (email !== authorEmail) {
+    throw new AppError(StatusCodes.UNAUTHORIZED, 'You are not authorized');
+  }
+
   const result = await Blog.findByIdAndUpdate(id, payload, { new: true });
   return result;
 };
@@ -23,7 +51,27 @@ const getAllBlogs = async (query: Record<string, unknown>) => {
   return result;
 };
 
-const deleteBlog = async (id: string) => {
+const deleteBlog = async (id: string, token: string) => {
+  const data = await Blog.findById(id);
+
+  const authorId = data?.author;
+  const authorEmail = (await User.findById(authorId))?.email;
+
+  if (!token) {
+    throw new AppError(StatusCodes.UNAUTHORIZED, 'You are not authorized');
+  }
+
+  const decoded = jwt.verify(
+    token,
+    config.jwt_access_secret as string,
+  ) as JwtPayload;
+
+  const { email } = decoded;
+
+  if (email !== authorEmail) {
+    throw new AppError(StatusCodes.UNAUTHORIZED, 'You are not authorized');
+  }
+
   const result = await Blog.findByIdAndDelete(id);
   return result;
 };
